requirements.txt
fastapi
uvicorn
nest-asyncio
pyngrok
numpy
pandas
scikit-learn
matplotlib
seaborn
gdown
requests
pillow
pydantic
